Add files via upload
The Smart Student Assistant System is a Python program that helps students manage their academic details. It allows users to store profile information, calculate attendance and marks, and maintain a to-do list. The project uses file handling to save data permanently and provides a simple menu-driven interface for easy use.